pyproject.toml
[tool.poetry]
name = "twitter-bitcoin-api"
version = "0.1.0"
description = ""
authors = ["odylomv <odylomv@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.11"
Flask = "^2.2.1"
tweepy = "^4.10.0"
Flask-Cors = "^3.0.10"
stegano = "^0.11.1"
steganography = "^0.1.1"
pycryptodome = "^3.16.0"
python-bitcoinrpc = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
